requirements.txt
# Core dependencies
pydantic>=2.5.2
python-dotenv>=1.0.0
requests>=2.31.0
ipywidgets>=7.7.1

# PDF processing
pypdf>=4.0.0
unstructured>=0.11.0
pdf2image>=1.16.3
pytesseract>=0.3.10

# Vector stores
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0

# CSV/JSON processing
pandas>=2.0.3

# Utilities
numpy>=1.24.3
tqdm>=4.66.1

# Google Colab specific
google-colab
ipython>=7.0.0
